feat: outputView 추가
