+++ src/main/java/baseball/Application.java
@@ -1,6 +1,7 @@
 package baseball;
 
 import baseball.adapter.in.game.ClientInputView;
+import baseball.adapter.in.game.ClientOutputView;
 import baseball.adapter.in.game.GameClient;
 import baseball.application.port.in.GameUseCase;
 import baseball.application.port.service.BaseballGame;
@@ -10,8 +11,13 @@ import baseball.domain.BaseballRandomDigitGenerator;
 public class Application {
 
     public static void main(String[] args) {
-        GameClient gameClient = new GameClient(generateGameUsecase(), generateClientInputView());
+        GameClient gameClient = new GameClient(
+                generateGameUsecase(), generateClientInputView(), generateClientOutputView());
         gameClient.playGame();
+    }
+
+    private static ClientOutputView generateClientOutputView() {
+        return new ClientOutputView();
     }
 
     private static GameUseCase generateGameUsecase() {
